vlm_ros/requirements.txt
numpy==1.23.4
omegaconf==2.3.0
pandas==1.5.3
peft==0.4.0
Pillow==9.4.0
PyYAML==6.0
scikit_learn==1.2.2
sconf==0.2.5
timm==0.9.8
tokenizers==0.13.3
tqdm==4.64.1
transformers==4.31.0
einops==0.7.0
einops-exts==0.0.4
prettytable==3.9.0
sentencepiece==0.1.99
shortuuid
accelerate==0.21.0
bitsandbytes==0.41.0
pydantic==1.8.2
requests==2.31.0
httpx==0.24.0
uvicorn
fastapi
decord==0.6.0
